% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add-quality-control-flags.R
\name{add_quality_control_flags}
\alias{add_quality_control_flags}
\title{Add quality control flags to each GBIF occurrence record}
\usage{
add_quality_control_flags(occurrence_data)
}
\arguments{
\item{occurrence_data}{a data frame of occurrence data extracted from GBIF}
}
\value{
a data frame with eight additional columns corresponding to the
different quality control flags
}
\description{
Check each GBIF occurrence record (i.e. row) for common quality control
concerns and assign records that exhibit a particular quality concern \code{TRUE}
for the given column, else \code{FALSE}.
}
