% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bryonia-alba.R
\docType{data}
\name{bryonia_alba}
\alias{bryonia_alba}
\title{GBIF occurrence data for Bryonia alba}
\format{
A data frame with 9,962 rows and 6 variables:
\describe{
\item{datasetName}{The name identifying the data set from which the record was derived.}
\item{decimalLatitude}{The geographic latitude (in decimal degrees, using the WGS84 datum) of the geographic centre of the location of the occurrence.}
\item{decimalLongitude}{The geographic longitude (in decimal degrees, using the WGS84 datum) of the geographic centre of the location of the occurrence.}
\item{habitat}{A category or description of the habitat in which the dwc:Event occurred.}
\item{locality}{The specific description of the place.}
\item{year}{The four-digit year in which the event occurred, according to the Common Era calendar.}
}
}
\source{
GBIF.org (22 July 2026) GBIF Occurrence Download https://doi.org/10.15468/dl.up8hzy
}
\usage{
bryonia_alba
}
\description{
Occurrence data for the plant Bryonia alba, the Eastern white-bryony, from
the Global Biodiversity Information Facility.
}
\keyword{datasets}
