% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-occurrence-data-required-cols.R
\name{check_occurrence_data_required_cols}
\alias{check_occurrence_data_required_cols}
\title{Check if the GBIF occurrence data has the required columns}
\usage{
check_occurrence_data_required_cols(occurrence_data)
}
\arguments{
\item{occurrence_data}{a dataframe of GBIF occurrence data}
}
\value{
returns invsibily if all required columns are present. If not, an
error is raised.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\details{
An internal helper to check if the GBIF occurrence data has the required
columns
}
\keyword{internal}
