% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_wracs_data_dir.R
\name{get_wracs_data_dir}
\alias{get_wracs_data_dir}
\title{Get the WRACS data directory}
\usage{
get_wracs_data_dir()
}
\value{
a path to the WRACS data directory
}
\description{
Get the WRACS data directory which will store rasters of derived climate
variables downloaded from Zenodo. The location of the WRACS data directory
can be customized by setting \code{wracs_data_dir} as an environment variable.
}
