% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_wracs_data_doi.R
\name{get_wracs_data_doi}
\alias{get_wracs_data_doi}
\title{Get the DOI for the specified version of the WRACS data}
\usage{
get_wracs_data_doi(wracs_data_version = NULL)
}
\arguments{
\item{wracs_data_version}{A positive integer indicating the Zenodo record version.}
}
\description{
Get the DOI for the specified version of the WRACS data
}
