% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_wracs_data_version.R
\name{set_wracs_data_version}
\alias{set_wracs_data_version}
\title{Set the WRACS data version}
\usage{
set_wracs_data_version(wracs_data_version)
}
\arguments{
\item{wracs_data_version}{A positive integer indicating the Zenodo record version.}
}
\description{
Set the version of WRACS data to download from Zenodo and use in climate
suitability analyses
}
