% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wracs-package.R
\docType{package}
\name{wracs-package}
\alias{wracs}
\alias{wracs-package}
\title{wracs: Assessing Climatic Suitability for Weed Risk Assessments}
\description{
Assess the climatic suitability of a weed species within North America on the basis of Koppen-Geiger classes, Plant Hardiness zones, and Precipitation Bands.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://cfia-pdam.github.io/wracs/}
}

}
\author{
\strong{Maintainer}: Josh Persi \email{josh.persi@inspection.gc.ca}

Authors:
\itemize{
  \item Josh Persi \email{josh.persi@inspection.gc.ca}
  \item Megan Abergel \email{megan.abergel@inspection.gc.ca}
}

}
\keyword{internal}
