% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add-covariates.R
\name{add_covariates}
\alias{add_covariates}
\title{Add climate data covariates}
\usage{
add_covariates(gbif_occurrence_data = NULL, covariate_raster = NULL)
}
\arguments{
\item{gbif_occurrence_data}{a dataframe of GBIF occurrence data}

\item{covariate_raster}{a terra raster object of covariate data}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
