% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_covariates.R
\name{add_covariates}
\alias{add_covariates}
\title{Add climate data covariates}
\usage{
add_covariates(
  gbif_occurrence_data = NULL,
  emission_scenario = base::c("ssp370"),
  time_period = base::c("2011.2040", "2041.2070")
)
}
\description{
Add climate data covariates
}
