% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bryonia_alba.R
\docType{data}
\name{bryonia_alba}
\alias{bryonia_alba}
\title{GBIF occurrence data for Bryonia alba}
\format{
A data frame with 9,913 rows and 230 variables:
\describe{
\item{gbifID}{Unique GBIF key for the occurrence. We aim to keep these keys stable, but this is not possible in every case.}
\item{accessRights}{Information about who can access the resource or an indication of its security status.}
\item{bibliographicCitation}{A bibliographic reference for the resource.}
\item{language}{A language of the resource.}
\item{license}{A legal document giving official permission to do something with the occurrence.}
\item{modified}{The most recent date-time on which the occurrence was changed, according to the publisher.}
\item{publisher}{The name of the organization publishing this record}
\item{references}{A related resource that is referenced, cited, or otherwise pointed to by the described resource.}
\item{rightsHolder}{A person or organization owning or managing rights over the resource.}
\item{type}{The nature or genre of the resource.}
\item{institutionID}{An identifier for the institution having custody of the object(s) or information referred to in the record.}
\item{collectionID}{An identifier for the collection or dataset from which the record was derived.}
\item{datasetID}{An identifier for the set of data. May be a global unique identifier or an identifier specific to a collection or institution.}
\item{institutionCode}{The name (or acronym) in use by the institution having custody of the object(s) or information referred to in the record.}
\item{collectionCode}{The name, acronym, coden, or initialism identifying the collection or data set from which the record was derived.}
\item{datasetName}{The name identifying the data set from which the record was derived.}
\item{ownerInstitutionCode}{The name (or acronym) in use by the institution having ownership of the object(s) or information referred to in the record.}
\item{basisOfRecord}{The values of the Darwin Core term Basis of Record which can apply to occurrences. See GBIF’s Darwin Core Type Vocabulary for definitions.}
\item{informationWithheld}{Additional information that exists, but that has not been shared in the given record.}
\item{dataGeneralizations}{Actions taken to make the shared data less specific or complete than in its original form. Suggests that alternative data of higher quality may be available on request.}
\item{dynamicProperties}{A list of additional measurements, facts, characteristics, or assertions about the record. Meant to provide a mechanism for structured content.}
\item{occurrenceID}{An identifier for the dwc:Occurrence (as opposed to a particular digital record of the dwc:Occurrence). In the absence of a persistent global unique identifier, construct one from a combination of identifiers in the record that will most closely make the dwc:occurrenceID globally unique.}
\item{catalogNumber}{An identifier (preferably unique) for the record within the data set or collection.}
\item{recordNumber}{An identifier given to the dwc:Occurrence at the time it was recorded. Often serves as a link between field notes and a dwc:Occurrence record, such as a specimen collector’s number.}
\item{recordedBy}{A person, group, or organization responsible for recording the original occurrence.}
\item{recordedByID}{A list (concatenated and separated) of the globally unique identifier for the person, people, groups, or organizations responsible for recording the original dwc:Occurrence.}
\item{individualCount}{The number of individuals present at the time of the Occurrence.}
\item{organismQuantity}{A number or enumeration value for the quantity of organisms.}
\item{organismQuantityType}{The type of quantification system used for the quantity of organisms.}
\item{sex}{The sex of the biological individual(s) represented in the dwc:Occurrence.}
\item{lifeStage}{The age class or life stage of the Organism(s) at the time the Occurrence was recorded. Values are aligned to the GBIF LifeStage vocabulary}
\item{reproductiveCondition}{The reproductive condition of the biological individual(s) represented in the dwc:Occurrence.}
\item{caste}{Categorisation of individuals for eusocial species (including some mammals and arthropods).}
\item{behavior}{The behavior shown by the subject at the time the dwc:Occurrence was recorded.}
\item{vitality}{An indication of whether a dwc:Organism was alive or dead at the time of collection or observation.}
\item{establishmentMeans}{Statement about whether an organism or organisms have been introduced to a given place and time through the direct or indirect activity of modern humans. Values are aligned to the GBIF EstablishmentMeans vocabulary, which is derived from the Darwin Core EstablishmentMeans vocabulary.}
\item{degreeOfEstablishment}{The degree to which an Organism survives, reproduces, and expands its range at the given place and time. Values are aligned to the GBIF DegreeOfEstablishment vocabulary, which is derived from the Darwin Core DegreeOfEstablishment vocabulary.}
\item{pathway}{The process by which an Organism came to be in a given place at a given time. Values are aligned to the GBIF Pathway vocabulary, which is derived from the Darwin Core Pathway vocabulary.}
\item{georeferenceVerificationStatus}{A categorical description of the extent to which the georeference has been verified to represent the best possible spatial description for the dcterms:Location of the dwc:Occurrence.}
\item{occurrenceStatus}{A statement about the presence or absence of a Taxon at a Location. For definitions, see the GBIF occurrence status vocabulary.}
\item{preparations}{A preparation or preservation method for a specimen.}
\item{disposition}{The current state of a dwc:MaterialEntity with respect to a collection.}
\item{associatedOccurrences}{A list (concatenated and separated) of identifiers of other dwc:Occurrence records and their associations to this dwc:Occurrence.}
\item{associatedReferences}{A list (concatenated and separated) of identifiers (publication, bibliographic reference, global unique identifier, URI) of literature associated with the dwc:Occurrence.}
\item{associatedSequences}{A list (concatenated and separated) of identifiers (publication, global unique identifier, URI) of genetic sequence information associated with the material entity.}
\item{associatedTaxa}{A list (concatenated and separated) of identifiers or names of dwc:Taxon records and the associations of this dwc:Occurrence to each of them.}
\item{otherCatalogNumbers}{A list (concatenated and separated) of previous or alternate fully qualified catalogue numbers or other human-used identifiers for the same occurrence, whether in the current or any other data set or collection.}
\item{occurrenceRemarks}{Comments or notes about the dwc:Occurrence.}
\item{organismID}{An identifier for the dwc:Organism instance (as opposed to a particular digital record of the dwc:Organism). May be a globally unique identifier or an identifier specific to the data set.}
\item{organismName}{A textual name or label assigned to a dwc:Organism instance.}
\item{organismScope}{A description of the kind of dwc:Organism instance. Can be used to indicate whether the dwc:Organism instance represents a discrete organism or if it represents a particular type of aggregation.}
\item{associatedOrganisms}{A list (concatenated and separated) of identifiers of other dwc:Organisms and the associations of this dwc:Organism to each of them.}
\item{previousIdentifications}{A list (concatenated and separated) of previous assignments of names to the dwc:Organism.}
\item{organismRemarks}{Comments or notes about the dwc:Organism instance.}
\item{materialEntityID}{An identifier for a particular instance of a dwc:MaterialEntity.}
\item{materialEntityRemarks}{Comments or notes about the dwc:MaterialEntity instance.}
\item{verbatimLabel}{The content of this term should include no embellishments, prefixes, headers or other additions made to the text. Abbreviations must not be expanded and supposed misspellings must not be corrected. Lines or breakpoints between blocks of text that could be verified by seeing the original labels or images of them may be used. Examples of material entities include preserved specimens, fossil specimens, and material samples. Best practice is to use UTF-8 for all characters. Best practice is to add comment “verbatimLabel derived from human transcription” in dwc:occurrenceRemarks.}
\item{materialSampleID}{An identifier for the dwc:MaterialSample (as opposed to a particular digital record of the dwc:MaterialSample). In the absence of a persistent global unique identifier, construct one from a combination of identifiers in the record that will most closely make the dwc:materialSampleID globally unique.}
\item{eventID}{An identifier for the set of information associated with a dwc:Event (something that occurs at a place and time). May be a global unique identifier or an identifier specific to the data set.}
\item{parentEventID}{An identifier for the broader dwc:Event that groups this and potentially other dwc:Events.}
\item{eventType}{The type for sampling event records.}
\item{fieldNumber}{An identifier given to the dwc:Event in the field. Often serves as a link between field notes and the dwc:Event.}
\item{eventDate}{The date-time during which an Event occurred. For occurrences, this is the date-time when the event was recorded. Not suitable for a time in a geological context.}
\item{eventTime}{The time or interval during which a dwc:Event occurred.}
\item{startDayOfYear}{The latest integer day of the year on which the Event occurred (1 for 1 January, 365 for 31 December, except in a leap year, in which case it is 366).}
\item{endDayOfYear}{The earliest integer day of the year on which the Event occurred (1 for 1 January, 365 for 31 December, except in a leap year, in which case it is 366).}
\item{year}{The four-digit year in which the event occurred, according to the Common Era calendar.}
\item{month}{The integer month in which the Event occurred.}
\item{day}{The integer day of the month on which the Event occurred.}
\item{verbatimEventDate}{The verbatim original representation of the date and time information for a dwc:Event.}
\item{habitat}{A category or description of the habitat in which the dwc:Event occurred.}
\item{samplingProtocol}{The methods or protocols used during an Event, denoted by an IRI.}
\item{sampleSizeValue}{A numeric value for a measurement of the size (time duration, length, area, or volume) of a sample in a sampling event.}
\item{sampleSizeUnit}{The unit of measurement of the size (time duration, length, area, or volume) of a sample in a sampling event.}
\item{samplingEffort}{The amount of effort expended during a dwc:Event.}
\item{fieldNotes}{One of a) an indicator of the existence of, b) a reference to (publication, URI), or c) the text of notes taken in the field about the dwc:Event.}
\item{eventRemarks}{Comments or notes about the dwc:Event.}
\item{locationID}{An identifier for the set of dcterms:Location information. May be a global unique identifier or an identifier specific to the data set.}
\item{higherGeographyID}{An identifier for the geographic region within which the dcterms:Location occurred.}
\item{higherGeography}{A list (concatenated and separated) of geographic names less specific than the information captured in the dwc:locality term.}
\item{continent}{The continent, based on a 7 continent model described on Wikipedia and the World Geographical Scheme for Recording Plant Distributions (WGSRPD). In particular this splits the Americas into North and South America with North America including the Caribbean (except Trinidad and Tobago) and reaching down and including Panama. See the GBIF Continents for the exact divisions. This is a geographical division.  See GBIFRegion for GBIF’s political divisions.}
\item{waterBody}{The name of the water body in which the Location occurs.}
\item{islandGroup}{The name of the island group in which the dcterms:Location occurs.}
\item{island}{The name of the island on or near which the dcterms:Location occurs.}
\item{countryCode}{The 2-letter country code (as per ISO-3166-1) of the country, territory or area in which the occurrence was recorded.}
\item{stateProvince}{The name of the next-smaller administrative region than country (state, province, canton, department, region, etc.) in which the occurrence occurs. This value is unaltered by GBIF’s processing; see also the GADM fields.}
\item{county}{The full, unabbreviated name of the next smaller administrative region than stateProvince (county, shire, department, etc.) in which the dcterms:Location occurs.}
\item{municipality}{The full, unabbreviated name of the next smaller administrative region than county (city, municipality, etc.) in which the dcterms:Location occurs. Do not use this term for a nearby named place that does not contain the actual dcterms:Location.}
\item{locality}{The specific description of the place.}
\item{verbatimLocality}{The original textual description of the place.}
\item{verbatimElevation}{The original description of the elevation (altitude, usually above sea level) of the Location.}
\item{verticalDatum}{The vertical datum used as the reference upon which the values in the elevation terms are based.}
\item{verbatimDepth}{The original description of the depth below the local surface.}
\item{minimumDistanceAboveSurfaceInMeters}{The lesser distance in a range of distance from a reference surface in the vertical direction, in meters. Use positive values for locations above the surface, negative values for locations below. If depth measures are given, the reference surface is the location given by the depth, otherwise the reference surface is the location given by the elevation.}
\item{maximumDistanceAboveSurfaceInMeters}{The greater distance in a range of distance from a reference surface in the vertical direction, in meters. Use positive values for locations above the surface, negative values for locations below. If depth measures are given, the reference surface is the location given by the depth, otherwise the reference surface is the location given by the elevation.}
\item{locationAccordingTo}{Information about the source of this dcterms:Location information. Could be a publication (gazetteer), institution, or team of individuals.}
\item{locationRemarks}{Comments or notes about the dcterms:Location.}
\item{decimalLatitude}{The geographic latitude (in decimal degrees, using the WGS84 datum) of the geographic centre of the location of the occurrence.}
\item{decimalLongitude}{The geographic longitude (in decimal degrees, using the WGS84 datum) of the geographic centre of the location of the occurrence.}
\item{coordinateUncertaintyInMeters}{The horizontal distance (in metres) from the given decimalLatitude and decimalLongitude describing the smallest circle containing the whole of the Location.}
\item{coordinatePrecision}{A decimal representation of the precision of the coordinates given in the decimalLatitude and decimalLongitude.}
\item{pointRadiusSpatialFit}{The ratio of the area of the point-radius (dwc:decimalLatitude, dwc:decimalLongitude, dwc:coordinateUncertaintyInMeters) to the area of the true (original, or most specific) spatial representation of the dcterms:Location. Legal values are 0, greater than or equal to 1, or undefined. A value of 1 is an exact match or 100\% overlap. A value of 0 should be used if the given point-radius does not completely contain the original representation. The dwc:pointRadiusSpatialFit is undefined (and should be left empty) if the original representation is any geometry without area (e.g., a point or polyline) and without uncertainty and the given georeference is not that same geometry (without uncertainty). If both the original and the given georeference are the same point, the dwc:pointRadiusSpatialFit is 1.}
\item{verbatimCoordinateSystem}{The coordinate format for the dwc:verbatimLatitude and dwc:verbatimLongitude or the dwc:verbatimCoordinates of the dcterms:Location.}
\item{verbatimSRS}{The ellipsoid, geodetic datum, or spatial reference system (SRS) upon which coordinates given in dwc:verbatimLatitude and dwc:verbatimLongitude, or dwc:verbatimCoordinates are based.}
\item{footprintWKT}{A Well-Known Text (WKT) representation of the shape (footprint, geometry) that defines the dcterms:Location. A dcterms:Location may have both a point-radius representation (see dwc:decimalLatitude) and a footprint representation, and they may differ from each other.}
\item{footprintSRS}{The ellipsoid, geodetic datum, or spatial reference system (SRS) upon which the geometry given in dwc:footprintWKT is based.}
\item{footprintSpatialFit}{The ratio of the area of the dwc:footprintWKT to the area of the true (original, or most specific) spatial representation of the dcterms:Location. Legal values are 0, greater than or equal to 1, or undefined. A value of 1 is an exact match or 100\% overlap. A value of 0 should be used if the given dwc:footprintWKT does not completely contain the original representation. The dwc:footprintSpatialFit is undefined (and should be left empty) if the original representation is any geometry without area (e.g., a point or polyline) and without uncertainty and the given georeference is not that same geometry (without uncertainty). If both the original and the given georeference are the same point, the dwc:footprintSpatialFit is 1.}
\item{georeferencedBy}{A list (concatenated and separated) of names of people, groups, or organizations who determined the georeference (spatial representation) for the dcterms:Location.}
\item{georeferencedDate}{The date on which the dcterms:Location was georeferenced.}
\item{georeferenceProtocol}{A description or reference to the methods used to determine the spatial footprint, coordinates, and uncertainties.}
\item{georeferenceSources}{A list (concatenated and separated) of maps, gazetteers, or other resources used to georeference the dcterms:Location, described specifically enough to allow anyone in the future to use the same resources.}
\item{georeferenceRemarks}{Comments or notes about the spatial description determination, explaining assumptions made in addition or opposition to the those formalized in the method referred to in dwc:georeferenceProtocol.}
\item{geologicalContextID}{An identifier for the set of information associated with a dwc:GeologicalContext (the location within a geological context, such as stratigraphy). May be a global unique identifier or an identifier specific to the data set.}
\item{earliestEonOrLowestEonothem}{The full name of the earliest possible geochronologic eon or lowest chrono-stratigraphic eonothem or the informal name ("Precambrian") attributable to the stratigraphic horizon from which the dwc:MaterialEntity was collected.}
\item{latestEonOrHighestEonothem}{The full name of the latest possible geochronologic eon or highest chrono-stratigraphic eonothem or the informal name ("Precambrian") attributable to the stratigraphic horizon from which the dwc:MaterialEntity was collected.}
\item{earliestEraOrLowestErathem}{The full name of the earliest possible geochronologic era or lowest chronostratigraphic erathem attributable to the stratigraphic horizon from which the dwc:MaterialEntity was collected.}
\item{latestEraOrHighestErathem}{The full name of the latest possible geochronologic era or highest chronostratigraphic erathem attributable to the stratigraphic horizon from which the dwc:MaterialEntity was collected.}
\item{earliestPeriodOrLowestSystem}{The full name of the earliest possible geochronologic period or lowest chronostratigraphic system attributable to the stratigraphic horizon from which the dwc:MaterialEntity was collected.}
\item{latestPeriodOrHighestSystem}{The full name of the latest possible geochronologic period or highest chronostratigraphic system attributable to the stratigraphic horizon from which the dwc:MaterialEntity was collected.}
\item{earliestEpochOrLowestSeries}{The full name of the earliest possible geochronologic epoch or lowest chronostratigraphic series attributable to the stratigraphic horizon from which the dwc:MaterialEntity was collected.}
\item{latestEpochOrHighestSeries}{The full name of the latest possible geochronologic epoch or highest chronostratigraphic series attributable to the stratigraphic horizon from which the dwc:MaterialEntity was collected.}
\item{earliestAgeOrLowestStage}{The full name of the earliest possible geochronologic age or lowest chronostratigraphic stage attributable to the stratigraphic horizon from which the dwc:MaterialEntity was collected.}
\item{latestAgeOrHighestStage}{The full name of the latest possible geochronologic age or highest chronostratigraphic stage attributable to the stratigraphic horizon from which the dwc:MaterialEntity was collected.}
\item{lowestBiostratigraphicZone}{The full name of the lowest possible geological biostratigraphic zone of the stratigraphic horizon from which the dwc:MaterialEntity was collected.}
\item{highestBiostratigraphicZone}{The full name of the highest possible geological biostratigraphic zone of the stratigraphic horizon from which the dwc:MaterialEntity was collected.}
\item{lithostratigraphicTerms}{The combination of all lithostratigraphic names for the rock from which the dwc:MaterialEntity was collected.}
\item{group}{The full name of the lithostratigraphic group from which the dwc:MaterialEntity was collected.}
\item{formation}{The full name of the lithostratigraphic formation from which the dwc:MaterialEntity was collected.}
\item{member}{The full name of the lithostratigraphic member from which the dwc:MaterialEntity was collected.}
\item{bed}{The full name of the lithostratigraphic bed from which the dwc:MaterialEntity was collected.}
\item{identificationID}{An identifier for the dwc:Identification (the body of information associated with the assignment of a scientific name). May be a global unique identifier or an identifier specific to the data set.}
\item{verbatimIdentification}{A string representing the taxonomic identification as it appeared in the original record.}
\item{identificationQualifier}{A brief phrase or a standard term ("cf.", "aff.") to express the determiner’s doubts about the dwc:Identification.}
\item{typeStatus}{A list (concatenated and separated) of nomenclatural types (type status, typified scientific name, publication) applied to the occurrence.}
\item{identifiedBy}{A list (concatenated and separated) of names of people, groups, or organizations who assigned the Taxon to the occurrence.}
\item{identifiedByID}{A list (concatenated and separated) of the globally unique identifier for the person, people, groups, or organizations responsible for assigning the dwc:Taxon to the subject.}
\item{dateIdentified}{The date on which the subject was determined as representing the Taxon.}
\item{identificationReferences}{A list (concatenated and separated) of references (publication, global unique identifier, URI) used in the dwc:Identification.}
\item{identificationVerificationStatus}{A categorical indicator of the extent to which the taxonomic identification has been verified to be correct.}
\item{identificationRemarks}{Comments or notes about the dwc:Identification.}
\item{taxonID}{An identifier for the set of dwc:Taxon information. May be a global unique identifier or an identifier specific to the data set.}
\item{scientificNameID}{An identifier for the nomenclatural (not taxonomic) details of a scientific name.}
\item{acceptedNameUsageID}{An identifier for the name usage (documented meaning of the name according to a source) of the currently valid (zoological) or accepted (botanical) taxon.}
\item{parentNameUsageID}{An identifier for the name usage (documented meaning of the name according to a source) of the direct, most proximate higher-rank parent taxon (in a classification) of the most specific element of the dwc:scientificName.}
\item{originalNameUsageID}{An identifier for the name usage (documented meaning of the name according to a source) in which the terminal element of the dwc:scientificName was originally established under the rules of the associated dwc:nomenclaturalCode.}
\item{nameAccordingToID}{An identifier for the source in which the specific taxon concept circumscription is defined or implied. See dwc:nameAccordingTo.}
\item{namePublishedInID}{An identifier for the publication in which the dwc:scientificName was originally established under the rules of the associated dwc:nomenclaturalCode.}
\item{taxonConceptID}{An identifier for the taxonomic concept to which the record refers - not for the nomenclatural details of a dwc:Taxon.}
\item{scientificName}{The scientific name (including authorship) for the taxon from the GBIF backbone matched to this occurrence.  This could be a synonym, see also acceptedScientificName.}
\item{acceptedNameUsage}{The full name, with authorship and date information if known, of the currently valid (zoological) or accepted (botanical) dwc:Taxon.}
\item{parentNameUsage}{The full name, with authorship and date information if known, of the direct, most proximate higher-rank parent dwc:Taxon (in a classification) of the most specific element of the dwc:scientificName.}
\item{originalNameUsage}{The taxon name, with authorship and date information if known, as it originally appeared when first established under the rules of the associated dwc:nomenclaturalCode. The basionym (botany) or basonym (bacteriology) of the dwc:scientificName or the senior/earlier homonym for replaced names.}
\item{nameAccordingTo}{The reference to the source in which the specific taxon concept circumscription is defined or implied - traditionally signified by the Latin "sensu" or "sec." (from secundum, meaning "according to"). For taxa that result from identifications, a reference to the keys, monographs, experts and other sources should be given.}
\item{namePublishedIn}{A reference for the publication in which the dwc:scientificName was originally established under the rules of the associated dwc:nomenclaturalCode.}
\item{namePublishedInYear}{The four-digit year in which the dwc:scientificName was published.}
\item{higherClassification}{A list (concatenated and separated) of taxa names terminating at the rank immediately superior to the referenced dwc:Taxon.}
\item{kingdom}{The kingdom name (excluding authorship) for the kingdom from the GBIF backbone matched to this occurrence.}
\item{phylum}{The phylum name (excluding authorship) for the phylum from the GBIF backbone matched to this occurrence.}
\item{class}{The class name (excluding authorship) for the class from the GBIF backbone matched to this occurrence.}
\item{order}{The order name (excluding authorship) for the order from the GBIF backbone matched to this occurrence.}
\item{superfamily}{The full scientific name of the superfamily in which the dwc:Taxon is classified.}
\item{family}{The family name (excluding authorship) for the family from the GBIF backbone matched to this occurrence.}
\item{subfamily}{The full scientific name of the subfamily in which the dwc:Taxon is classified.}
\item{tribe}{The full scientific name of the tribe in which the dwc:Taxon is classified.}
\item{subtribe}{The full scientific name of the subtribe in which the dwc:Taxon is classified.}
\item{genus}{The genus name (excluding authorship) for the genus from the GBIF backbone matched to this occurrence.}
\item{genericName}{The genus name part of the species name from the GBIF backbone matched to this occurrence.}
\item{subgenus}{The subgenus name (excluding authorship) for the subgenus from the GBIF backbone matched to this occurrence.}
\item{infragenericEpithet}{The infrageneric part of a binomial name at ranks above species but below genus.}
\item{specificEpithet}{The specific name part of the species name from the GBIF backbone matched to this occurrence.}
\item{infraspecificEpithet}{The infraspecific name part of the species name from the GBIF backbone matched to this occurrence.}
\item{cultivarEpithet}{Part of the name of a cultivar, cultivar group or grex that follows the dwc:scientificName.}
\item{taxonRank}{The taxonomic rank of the most specific name in the scientificName.}
\item{verbatimTaxonRank}{The taxonomic rank of the most specific name in the dwc:scientificName as it appears in the original record.}
\item{vernacularName}{A common or vernacular name.}
\item{nomenclaturalCode}{The nomenclatural code (or codes in the case of an ambiregnal name) under which the dwc:scientificName is constructed.}
\item{taxonomicStatus}{The status of the use of the scientificName as a label for a taxon.}
\item{nomenclaturalStatus}{The status related to the original publication of the name and its conformance to the relevant rules of nomenclature. It is based essentially on an algorithm according to the business rules of the code. It requires no taxonomic opinion.}
\item{taxonRemarks}{Comments or notes about the taxon or name.}
\item{datasetKey}{The UUID of the GBIF dataset containing this occurrence.}
\item{publishingCountry}{The country, territory or island based on ISO-3166 of the organization publishing the dataset containing this occurrence.}
\item{lastInterpreted}{The time this occurrence was last processed by GBIF’s interpretation system “Pipelines”. This is the time the record was last changed in GBIF, not the time the record was last changed by the publisher.  Data is also reprocessed when we changed the taxonomic backbone, geographic data sources or other interpretation procedures. An earlier interpretation system distinguished between “parsing” and “interpretation”, but in the current system there is only one process — the two dates will always be the same.}
\item{elevation}{Elevation (altitude) in metres above sea level.  This is not a current Darwin Core term.}
\item{elevationAccuracy}{The value of the potential error associated with the elevation.  This is not a current Darwin Core term.}
\item{depth}{Depth in metres below sea level.  This is not a current Darwin Core term.}
\item{depthAccuracy}{The value of the potential error associated with the depth.  This is not a current Darwin Core term.}
\item{distanceFromCentroidInMeters}{The distance in metres of the occurrence from a centroid known to be applied to occurrences during georeferencing.  This can potentially indicate low-precision georeferencing, check the values of coordinateUncertaintyInMeters and georeferenceRemarks.}
\item{issue}{A specific interpretation issue found during processing and interpretation of the record. See the list of occurrence issues and the OccurrenceIssue enumeration for possible values and definitions.}
\item{taxonomicIssue}{–}
\item{nonTaxonomicIssue}{–}
\item{mediaType}{The media type given as Dublin Core type values, in particular StillImage, MovingImage or Sound.}
\item{hasCoordinate}{Boolean indicating that a valid latitude and longitude exists.}
\item{hasGeospatialIssues}{Boolean indicating that some spatial validation rule has not passed.}
\item{taxonKey}{A taxon key from the GBIF backbone for the most specific (lowest rank) taxon for this occurrence.  This could be a synonym, see acceptedTaxonKey.}
\item{acceptedTaxonKey}{A taxon key from the GBIF backbone for the accepted taxon of this occurrence.}
\item{kingdomKey}{A taxon key from the GBIF backbone for the kingdom of thisoccurrence.}
\item{phylumKey}{A taxon key from the GBIF backbone for the phylum of thisoccurrence.}
\item{classKey}{A taxon key from the GBIF backbone for the class of thisoccurrence.}
\item{orderKey}{A taxon key from the GBIF backbone for the order of thisoccurrence.}
\item{superfamilyKey}{–}
\item{familyKey}{A taxon key from the GBIF backbone for the family of thisoccurrence.}
\item{subfamilyKey}{–}
\item{tribeKey}{–}
\item{subtribeKey}{–}
\item{genusKey}{A taxon key from the GBIF backbone for the genus of thisoccurrence.}
\item{subgenusKey}{A taxon key from the GBIF backbone for the subgenus of thisoccurrence.}
\item{speciesKey}{A taxon key from the GBIF backbone for the species of thisoccurrence.}
\item{species}{The species name (excluding authorship) for the species from the GBIF backbone matched to this occurrence.}
\item{acceptedScientificName}{The accepted scientific name (including authorship) for the taxon from the GBIF backbone matched to this occurrence.}
\item{verbatimScientificName}{Scientific name as provided by the source.}
\item{typifiedName}{The scientific name that is based on the type specimen. This is not yet a Darwin Core term, see the proposal to add it.}
\item{protocol}{The technical protocol by which this occurrence was retrieved from the publisher’s systems.}
\item{lastParsed}{The time this occurrence was last processed by GBIF’s interpretation system “Pipelines”. This is the time the record was last changed in GBIF, not the time the record was last changed by the publisher.  Data is also reprocessed when we changed the taxonomic backbone, geographic data sources or other interpretation procedures. An earlier interpretation system distinguished between “parsing” and “interpretation”, but in the current system there is only one process — the two dates will always be the same.}
\item{lastCrawled}{The time this occurrence was last retrieved from the publisher’s systems.}
\item{isInvasive}{–}
\item{repatriated}{Boolean indicating if the publishing country is different to the location country.}
\item{relativeOrganismQuantity}{The relative measurement of the quantity of the organism (i.e. without absolute units).}
\item{projectId}{Identifiers for projects related to this occurrence.}
\item{isSequenced}{Flag occurrence when associatedSequences/extension exists}
\item{gbifRegion}{An enumeration for all GBIF Regions. These are based on IPBES regions. This is a political division, part of GBIF’s governance structure.}
\item{publishedByGbifRegion}{An enumeration for all GBIF Regions. These are based on IPBES regions. This is a political division, part of GBIF’s governance structure.}
\item{level0Gid}{The identifier for the top-level division from the GADM database. This is usually a three-letter code from ISO 3166.}
\item{level0Name}{The English name for the top-level division from the GADM database.}
\item{level1Gid}{The identifier for the first-level division from the GADM database.}
\item{level1Name}{The English name for the first-level division from the GADM database.}
\item{level2Gid}{The identifier for the second-level division from the GADM database.}
\item{level2Name}{The English name for the second-level division from the GADM database.}
\item{level3Gid}{The identifier for the third-level division from the GADM database.}
\item{level3Name}{The English name for the third-level division from the GADM database.}
\item{iucnRedListCategory}{The IUCN Red List Category of the taxon of this occurrence. See the GBIF vocabulary for the values and their definitions, and the IUCN Red List of Threatened Species dataset in GBIF for the version of the Red List GBIF’s interpretation procedures are using.}
}
}
\source{
GBIF.org (30 June 2026) GBIF Occurrence Download https://doi.org/10.15468/dl.8frmpw
}
\usage{
bryonia_alba
}
\description{
Occurrence data for the plant Bryonia alba, the Eastern white-bryony, from
the Global Biodiversity Information Facility.
}
\keyword{datasets}
